requirements.txt
python-binance>=1.0.19
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0  # Technical analysis library
matplotlib>=3.7.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0 